# lib/api-zod/src/generated/types/currentAuthUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MAMA maternal access, monitoring and action API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface CurrentAuthUser {
  user: AuthUser | null;
}